Caching / Assertion improvements and fix
- Detection and elimination of similar operations was improved. Caching
is handled for other basic operations, e.g. calling a SHA256 gadget on
the same input wires will not yield any more constraints.

- This update fixes one case with AssertionBasicOp that was not handled
correctly. In the previous version, AssertionBasicOp  instructions were
not assumed to be cached, but the exclusion condition was not specified
properly, so they were dealt with as MultiplicationBasicOp.  Due to
this, an assertion pair like [w1*w2=w3,  w1*w2=w4], was considered as
one AssertionBasicOp  and the 2nd op would not appear in the circuit.
Same for the pair  w1*1=w2, w1*1=w3. This makes sense for
MultiplicationBasicOp when new wires are produced, but not for
AssertionBasicOp where all the inputs and output are existing wires.

This case does not appear in any of the gadgets, as they use assertions
infrequently in a way that does not result in the above cases.