Override hashCode() in ConstantMulBasicOp
ConstantMulBasicOp is different from the rest of the Basic Operations,
so hashCode() is overriden for consistency.
The previous version of ConstantMulBasicOp should not cause correctness
or performance (circuit size) issues in jsnark itself, but the update
helps avoid an inconsistency with xjsnark's optimizer which builds on
top of the basic classes.