Update LongIntegerModGadget
Change the behavior of LongIntegerModGadget to suit other use cases. Previously, this gadget was written mainly to support the RSA gadgets in this project. It made a restrictive assumption that the value of the second operand (the modulus) will be in a certain range. However, in the general case where the second operand can have a wider range of values, the older version of this gadget will not be able to support it (will give exceptions during evaluation as it was restrictive).

Note: The problem in the previous version of this gadget is not a security issue. Also, the efficiency of the existing RSA gadgets is the same before and after the change. This update fixes the gadget in order to support other use cases beyond the RSA examples in this project.