Fix for a special case in LongELement
This is related to issue 30 in xJsnark https://github.com/akosba/xjsnark/issues/30, in which an ArrayIndexOutOfBounds Exception was reported. The exception happens when the grouping optimization used for asserting the equality of two long elements results only in one group. In this case, some of the steps are unnecessary and lead to the exception.